WBackendVlc: Update OpenGL functions to integer

+++ src/SkMedia/src/media/WBackendVlc_p.h
@@ -154,10 +154,10 @@ public: // Variables
 
     QRect targetRect;
 
-    GLfloat targetX;
-    GLfloat targetY;
-    GLfloat targetWidth;
-    GLfloat targetHeight;
+    int targetX;
+    int targetY;
+    int targetWidth;
+    int targetHeight;
 
     bool shader;
